requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
python-multipart>=0.0.5

# Image processing
Pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.20.0

# Video processing
moviepy>=1.0.3
imageio>=2.25.0
imageio-ffmpeg>=0.4.8

# AI/ML libraries
openai>=1.0.0
requests>=2.25.0
httpx>=0.24.0
google-genai>=0.1.0

# Data processing
pyyaml>=6.0
jsonschema>=4.0.0
pandas>=1.3.0

# Storage and database
sqlalchemy>=1.4.0
alembic>=1.8.0
redis>=4.0.0
boto3>=1.26.0

# Utilities
python-dotenv>=0.19.0
loguru>=0.6.0
click>=8.0.0
rich>=12.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0

# Development
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0
